requirements.txt
pydantic==2.6.4
sqlmodel==0.0.16
requests==2.31.0
python-dotenv==1.0.1
polars==0.20.17
pymysql==1.1.0